libutee: TEE_MACCompareFinal(): allow shorter macLen
The caller of TEE_MACCompareFinal() might want to check only the
beginning of the MAC. This use case is prompted by some NIST test
vectors, for example in the file CMACGenAES256.rsp (contained
in cmactestvectors.zip [1] available from the Cryptographic Algorithm
Validation Program page [2]). For these vectors, the Tlen parameter
gives the size of the MAC to be taken into account.

Update TEE_MACCompareFinal() accordingly.

Link: [1] https://csrc.nist.gov/CSRC/media/Projects/Cryptographic-Algorithm-Validation-Program/documents/mac/cmactestvectors.zip
Link: [2] https://csrc.nist.gov/Projects/Cryptographic-Algorithm-Validation-Program/CAVP-TESTING-BLOCK-CIPHER-MODES
Signed-off-by: Jerome Forissier <jerome@forissier.org>