mk: LIBGCC_LOCATE_CFLAGS allows setting sysroot path for GCC
OP-TEE OS compilation fails with Yocto like build environment
as libgcc.a cannot be find when gcc.mk is executed [1]. A trick
used  was to set comp-cflagscore='--sysroot=/path/to/sysroot'
at OP-TEE OS build. This is no more sufficient as CXXFLAGS
(comp-cxxflagscore=...) may also be needed.

This change implements LIBGCC_LOCATE_CFLAGS as a better way to
path sysroot info for the expected purpose, as proposed in [2] and
[3], recently mentioned in [1] and used in several distributions.

Build environment are expected to set path when needed:
LIBGCC_LOCATE_CFLAGS='--sysroot=/path/to/sysroot'

Link: [1] https://github.com/OP-TEE/optee_os/issues/4673
Link: [2] https://github.com/OP-TEE/optee_os/pull/2853
Link: [3] https://github.com/OP-TEE/optee_os/issues/4485#issuecomment-803025139

Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>