core: lpae: support user mapping when base level is 0
User mapping (i.e. TAs) is expected to be at level 2,
so an level 1 entry points to this mapping.
If base level is 1, as it was supported, nothing changes.
If base level is 0 then an extra page is created at level 1,
so user mapping can be pointed from level 1 entry, as it's
supported by user mappings.

Signed-off-by: Izik Dubnov <izik@amazon.com>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>
Reviewed-by: Etienne Carriere <etienne.carriere@linaro.org>