core: add run-time integrity check service
Adds a run-time integrity check service as SMC fast calls available to
normal world.

Normal world can register physical memory which will be hashed in secure
world using the OPTEE_SMC_ADD_RTI_CHECK SMC function. This hash which
also is saved in secure world can later be checked using the
OPTEE_SMC_RUN_RTI_CHECK SMC function. For now only a debug print is
emitted in case of a mismatch of hash, but that is expected to be
changed in future patches.

A new enum teecore_memtypes, MEM_AREA_RTI_CHECK_VASPACE, is added to to
represent the new temporary memory mapping needed when calculating the
hashes.

The mappings are performed as fast calls and potentially also from a
secure interrupt handler so ideally no memory allocations should be
performed during that execution. The is solved by having a special
translation table reserved for these RTI check mappings.

Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>