[RFC 2] core: pta: add remote attestation PTA
Add a PTA to perform remote attestation of user space TAs. Enabled with
CFG_ATTESTATION_PTA=y.

This feature allows a user space TA to request a measurement of itself
from the OP-TEE core. Typical use case: a client application invokes
its companion TA to obtain secure services and it needs some guarantees
that the code running in the TEE is indeed the expected one and has not
been tampered with or corrupted. The CA generates a random nonce and
sends it to the TA using a custom command. The TA opens a session to
the attestation PTA and passes the nonce. The PTA measures the caller,
thus generating a hash, which is then signed together with the nonce.
At this point the CA (or any other entity involved in the call flow,
such as a remote server for instance) can verify the signature. If it
is valid and the nonce is the expected one, then the TA hash can be
trusted and compared against a known good value.

TODO: implement the nonce and signature.

Signed-off-by: Jerome Forissier <jerome@forissier.org>