[RFC v3] core: pta: add remote attestation PTA
Add a PTA to perform remote attestation of user space TAs. Enabled with
CFG_ATTESTATION_PTA=y.

This feature allows a client application to request a measurement of any
Trusted Application or TA shared library installed on the device. The CA
generates a random nonce and sends it to the attestation Pseudo TA,
along with the UUID of the TA to be measured. The TA hash is the value
found in the TA header (struct shdr::hash). It is signed together with
the nonce using a private key embedded into the PTA and provided by
CG_ATTESTATION_PTA_SIGN_KEY. The TA hash and the signature are both
returned to the caller.

Signed-off-by: Jerome Forissier <jerome@forissier.org>