[RFC] core: pta: add remote attestation PTA
Add a PTA to perform remote attestation of user space TAs. Enabled with
CFG_ATTESTATION_PTA=y.

This feature allows a client application to request a measurement of a
TA running in secure user space before invoking a command. The
measurement is an SHA-256 hash of the virtual memory pages that are
specific to the TA and are non writable -- basically, code and read-only
data of the main application, as well as any shared library in use when
the measurement is made.

This is work in progress. TBD:
- The current code returns a simple hash; adding some kind of digital
signature is needed.
- The TA to be measured is identified by a session ID, which does not
provide any protection against a CA asking for a hash of an unrelated TA
(a CA could guess a session ID used by another CA/TA pair). Perhaps not
a big issue, but anyways I will propose another RFC that addresses this
point.

Signed-off-by: Jerome Forissier <jerome@forissier.org>