[RFC v4] core: pta: add remote attestation PTA
Add a PTA to perform remote attestation of user space TAs. Enabled with
CFG_ATTESTATION_PTA=y.

This feature allows a client application to request a measurement of any
Trusted Application or TA shared library installed on the device. The CA
generates a random nonce and sends it to the attestation Pseudo TA,
along with the UUID of the TA to be measured. The measurement is the TA
hash found in the TA header (struct shdr::hash). It is hashed once more
together with the nonce, then the resulting digest is signed using a
private key that is generated by the PTA upon the first invocation. The
key is persistent (written into secure storage). A command can be used
by the client to retrieve the public key; the private key never leaves
the TEE. This command would typically be invoked only during device
provisioning so that the key pair is generated and the public key is
returned in a trusted environment.

Signed-off-by: Jerome Forissier <jerome@forissier.org>