drivers: stm32_bsec3: add BSEC driver support
Add the BSEC3 driver for STM32MP25x support.

The OTPs are shadowed in SRAM1, accesisible by M33 secure.

Each OTPs can be organized in NVMEM cells, with properties:

- "st,secure-otp": restrict the read of the secure OTPs,
  only accessible by TA and for the TDCID; its value is not
  shadowed and not available from non-secure.

- "st,non-secure-provisioning-otp": access granted to non-secure
  world only if the fuses are not locked.

If the shadow region (SRAM1 for example in STANDBY or LPLV-Stop)
is not preserved, it must be restored in pm resume function.

Signed-off-by: Lionel Debieve <lionel.debieve@foss.st.com>
Signed-off-by: Patrick Delaunay <patrick.delaunay@foss.st.com>
Signed-off-by: Thomas Bourgoin <thomas.bourgoin@foss.st.com>