scripts: add print_tee_hash.py
Adds a Python script to return the SHA256 hash of the .text* and .rodata*
sections of tee.elf, excluding the small area in .text that contains
data initialized early after boot. This hash value can later be compared
with the hash returned at runtime by the attestation PTA which will come
in the next commit.

Signed-off-by: Jerome Forissier <jerome@forissier.org>
Reviewed-by: Sumit Garg <sumit.garg@linaro.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>