ldelf: allow ELFs without DT_HASH to be loaded
It is possible for an ELF to be compiled with --hash-style=gnu;
some distros even have that as their default. This produces an ELF
with no .hash section, but with a .gnu.hash section in its stead.

As a first step towards detaching table-traversal code from
the rest of the symbol resolution code, do not fail in the case
of a missing DT_HASH, but use the old bruteforce lookup code
when elf->hashtab is NULL.

Signed-off-by: Elvira Khabirova <e.khabirova@omp.ru>