core: FF-A: Allow empty base-address in manifest
The base-address element in the manifest is optional. When no
base-address is found, the SPMC needs to allocate memory for the SP. For
each empty base-address, we increase the size of the fdt. This is needed
to make sure that the fdt hjas enough space to store the extra fields.

Signed-off-by: Jelle Sels <jelle.sels@arm.com>