core: crypto: sm2: remove TEE_ATTR_ECC_CURVE attribute from SM2 keys
SM2 key objects should not have a TEE_ATTR_ECC_CURVE attribute. For
example, the GP specification states that TEE_GenerateKey() can be
called with no attribute for the various SM2 algorithms. The current
implementation uses the same attribute templates than generic ECC
algorithms, which is wrong. Define specific variants for SM2 so that
TEE_ATTR_ECC_CURVE is not required when generating or populating SM2
keys and is not output either by TEE_GetObjectBufferAttribute().

Signed-off-by: Jerome Forissier <jerome@forissier.org>
Reviewed-by: Etienne Carriere <etienne.carriere@linaro.org>